prog 3 pseudo and code

+++ compute_sum_two_number/compute_sum_two_number.py
@@ -1,7 +1,9 @@
 # ask user to input two numbers
-input_number_one = float(input("Input First Number:"))
-input_number_two = float(input("Input Second Number:"))
+input_number_one = int(input("Input First Number:"))
+input_number_two = int(input("Input Second Number:"))
+
 # calculate the sum of inputted numbers
-sum = input_number_one+input_number_two
+sum = input_number_one + input_number_two
+
 # print sum of inputted numbers
 print(sum)
